requirements.txt
langchain
google-cloud-aiplatform>=1.56.0
google-cloud-discoveryengine
langchain-google-vertexai
streamlit
langchain_google_genai
streamlit-lottie
streamlit-authenticator
streamlit-on-Hover-tabs
google-genai
vertexai